Add files via upload
For this assignment I used tools like chatgpt in order to debugg the code, fix mistakes, and to improve the overall quality. OpenAI. (2024, November 30). ChatGPT: Assistance with debugging, enhancing functionality, and optimizing game code. OpenAI. https://chat.openai.com/. I attemted to do task 1-2-3-5 and to a certain extent this did not go so well.